pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mubashar-Bashir", email = "mubasharbashir003@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.59.8",
    "python-dotenv>=1.0.1",
]

[project.scripts]
myproject = "myproject:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
